move files to follow react file structure recommendation
