+++ src/components/projects/Entry.scss
