+++ src/components/resumes/Entry.jsx
@@ -1,48 +1,11 @@
-import * as Yup from "yup";
-import { postResume } from "./ResumeCreationApi.js";
+import { entry as entryValidations } from "./helpers/validations";
+import { postResume } from "./ResumesAPI.js";
 import { useFormik } from "formik";
 import { useHistory } from "react-router-dom";
 import { Button, TextField } from "@material-ui/core";
-import "./ResumeCreation.css";
+import "./Entry.css";
 
-const validationSchema = Yup.object({
-  birthdate: Yup.string()
-    .matches(
-      /^([0-2][0-9]|(3)[0-1])(\/)(((0)[0-9])|((1)[0-2]))(\/)\d{4}$/,
-      "Please enter valid date of birth"
-    )
-    .required("Date of birth field is required"),
-  direction: Yup.string()
-    .min(3, "Address is too short - should be 3 characters minimum")
-    .max(20, "Please enter an address with 20 characters or less")
-    .required("Address field is required"),
-  email: Yup.string()
-    .email("Invalid email address")
-    .required("Email field is required"),
-  firstName: Yup.string()
-    .min(3, "Name is too short - should be 3 characters minimum")
-    .matches(/^[A-Za-z ]*$/, "Please enter valid name")
-    .max(20, "Please enter a name with 20 characters or less")
-    .required("Name field is required"),
-  lastName: Yup.string()
-    .min(3, "Lastname is too short - should be 3 characters minimum")
-    .matches(/^[A-Za-z ]*$/, "Please enter valid lastname")
-    .max(20, "Please enter a lastname with 20 characters or less")
-    .required("Lastname field is required"),
-  phone: Yup.string()
-    .matches(/^[67]\d{7}$/g, "Invalid phone number")
-    .required("Phone number field is required"),
-  summary: Yup.string()
-    .min(20, "Bio description is too short - should be 20 characters minimum")
-    .matches(
-      /^[A-Za-z  .,']*$/g,
-      "Numbers and special characters are not allowed @()-*$#!+=^&"
-    )
-    .max(150, "Please enter a summary with 150 characters or less")
-    .required("Bio description field is required"),
-});
-
-function ResumeCreation() {
+function Entry() {
   const history = useHistory();
   const formik = useFormik({
     initialValues: {
@@ -57,14 +20,14 @@ function ResumeCreation() {
     onSubmit: async (values) => {
       const result = await postResume(values);
       if (!result.ok) {
-        history.push("/resumecreation");
+        history.push("/resumes/entry");
       } else {
         alert("Success! Welcome to TeamUp!");
         history.push("/");
         return result;
       }
     },
-    validationSchema,
+    validationSchema: entryValidations(),
   });
 
   return (
@@ -213,4 +176,4 @@ function ResumeCreation() {
   );
 }
 
-export default ResumeCreation;
+export default Entry;
